jira_automation/requirements.txt
atlassian-python-api==3.41.3
requests>=2.25.1
python-dotenv>=0.19.0